Extract shared per-backend benchmark boilerplate into bench_common.h
Moves BenchBackend, the HDF5/Memory URI builder, and pulse-open lifecycle
setup out of bench_equilibrium_read.cpp into bench_common.h so the three
follow-on benchmarks (issue #57) can reuse them instead of re-deriving the
same boilerplate. No change to the existing whole-IDS-read benchmark's
behavior or reported numbers (issue #58).
